backend/requirements.txt
fastapi==0.111.1
uvicorn==0.24.0
sqlalchemy==2.0.20
sse-starlette==1.6.1
python-dotenv==1.0.0
httpx==0.24.1
psycopg2-binary==2.9.6